backend/requirements.txt
flask
flask-cors
numpy
opencv-python
scipy
Pillow
